Refactor summariseGames to be used across processors
