+++ src/processors/boardgameSummariesByTags.js
@@ -1,4 +1,5 @@
 const reduceNameToBoardGameApiId = require('../util/reduceNameToBoardGameApiId')
+const summariseGames = require('../util/summariseGames')
 
 const log = []
 const playRecordsWithMoreThanOneGameFamily = []
@@ -19,10 +20,10 @@ function findGameFamily ({ playRecords }) {
 async function createBoardGameSummariesByTags (model) {
   const { index } = model.calisaurus
 
-  report('TODO: Time to get tagging?')
   report('Index:', Object.keys(index).join(', '))
   report('Love Letter:', JSON.stringify(index['love-letter'], null, 2))
 
+  report('Currently supported tags: gameFamily')
   const gameFamilies = Object.entries(index).reduce((acc, [name, gameData]) => {
     const gameFamily = findGameFamily(gameData)
     const gameFamilyId = reduceNameToBoardGameApiId(gameFamily)
@@ -39,6 +40,28 @@ async function createBoardGameSummariesByTags (model) {
     acc[gameFamilyId] = entry
     return acc
   }, {})
+
+  Object.entries(gameFamilies).forEach(([gameFamilyId, entry]) => {
+    const { playRecords } = entry
+    const timesInUse = playRecords.map(n => new Date(n.date).getTime())
+
+    const earliestTime = Math.min(...timesInUse)
+    const latestTime = Math.max(...timesInUse)
+
+    const earliestDate = new Date(earliestTime)
+    const latestDate = new Date(latestTime)
+
+    const stats = summariseGames({
+      games: playRecords,
+      startDate: earliestDate,
+      endDate: latestDate
+    })
+
+    entry.earliestDate = earliestDate.toISOString().substring(0, 10)
+    entry.latestDate = latestDate.toISOString().substring(0, 10)
+
+    Object.assign(entry, stats)
+  })
 
   const exampleDataStructure = {
     type: {
